online-course-platform/requirements.txt
streamlit
stripe
python-dotenv
